+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `ErasedSet::shrink_to`
+
 ## [0.6.0] - 2022-05-11
 
 ### Changed